requirements.txt
fastapi
uvicorn
pydantic
groq
python-pptx
reportlab
python-multipart
python-dotenv
httpx